+++ templates/template-flutter/lib/app/search/presenter/stores/search_store.dart
